// src/store/enums/CategoryActionsEnum.ts
export enum CategoryActionTypes {
  GET_CATEGORIES = "GET_CATEGORIES",
  GET_CATEGORIES_SUCCESS = "GET_CATEGORIES_SUCCESS",
  GET_CATEGORIES_FAILURE = "GET_CATEGORIES_FAILURE",
  CREATE_CATEGORY = "CREATE_CATEGORY",
  CREATE_CATEGORY_SUCCESS = "CREATE_CATEGORY_SUCCESS",
  CREATE_CATEGORY_FAILURE = "CREATE_CATEGORY_FAILURE",
  UPDATE_CATEGORY = "UPDATE_CATEGORY",
  UPDATE_CATEGORY_SUCCESS = "UPDATE_CATEGORY_SUCCESS",
  UPDATE_CATEGORY_FAILURE = "UPDATE_CATEGORY_FAILURE",
  DELETE_CATEGORY = "DELETE_CATEGORY",
  DELETE_CATEGORY_SUCCESS = "DELETE_CATEGORY_SUCCESS",
  DELETE_CATEGORY_FAILURE = "DELETE_CATEGORY_FAILURE",
}